feat: add register_info API and comprehensive test suite
- Add new register_info.hpp header with compile-time ISA detection
  - Enum instruction_set with 12 ISA variants (SSE2, AVX, ARM, PowerPC, MIPS)
  - Struct register_info with register counts and vector dimensions
  - Public API: detected_isa(), available_registers(), registers_for(isa)
  - Compile-time constants: fp_register_count(), vector_width_bits(), etc.
  - Register information for all major ISAs

- Update benchmarks to use xorshift32 and integer arithmetic
  - Replaced splitmix64 with xorshift32 for equal GCC/Clang vectorization
  - Converted from floating-point FMA chains to simple integer accumulation
  - Improved fairness of benchmark comparisons across compilers

- Add comprehensive register_info tests
  - 21 test cases covering all ISAs and API functions
  - Tests for compile-time constants and register properties
  - Runs across C++17, C++20, and C++23 standards

- Fix clang-tidy warnings in register_info.hpp
  - Use unsigned char base type for enum
  - Consolidate identical switch cases (AVX/AVX2, ARM SVE/SVE2)

+++ benchmarks/static_for_bench_dependent.cpp
@@ -2,7 +2,7 @@
 
 namespace poet_bench::static_for_support {
 
-void run_static_for_bench_dependent(std::uint64_t s) {
+void run_static_for_bench_dependent(std::uint32_t s) {
     ankerl::nanobench::Bench bench;
     bench.title("1. Compile-time dependent operations (different constants per iteration)");
     bench.minEpochTime(std::chrono::milliseconds{ 10 });
